+++ packages/contracts/src/support/env/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Support Env identifier
+ *
+ * @type {Symbol}
+ */
+export const SUPPORT_ENV: unique symbol = Symbol('@aedart/contracts/support/env');
+
+import Repository from "./Repository";
+export {
+    type Repository
+}