+++ shared/tests/webpack.config.js
@@ -1,4 +1,6 @@
-const fs = require('fs');
+const fs = require('node:fs');
+const dotEnv = require('dotenv');
+const webpack = require('webpack');
 
 /*****************************************************************
  * Load babel presents...
@@ -8,6 +10,16 @@ const babel = JSON.parse(
     fs.readFileSync('babel.config.json', 'utf8')
 );
 console.debug('Babel presets', babel.presets);
+
+/*****************************************************************
+ * Load .env content
+ ****************************************************************/
+
+// @see https://github.com/motdotla/dotenv?tab=readme-ov-file#parsing
+const ENV = dotEnv.parse(
+    fs.readFileSync('.env', 'utf8')
+);
+// console.debug('.env', ENV);
 
 /*****************************************************************
  * Webpack Configuration (for tests)
@@ -27,9 +39,18 @@ module.exports = {
                 ],
                 loader: "babel-loader",
                 options: {
-                    presets : babel.presets
+                    presets : babel.presets,
+                    plugins: babel.plugins
                 }
             }
-        ]
-    }
+        ],
+    },
+    plugins: [
+        
+        // Inject environment variables into token
+        // @see https://webpack.js.org/plugins/define-plugin/
+        new webpack.DefinePlugin({
+            __ENV__: JSON.stringify(ENV)
+        })
+    ]
 };